Break up DNS tests into 3 files to help with GH UI and IDE issues.
